added it to the save function
