+++ src/django/VLE/utils/generic_utils.py
@@ -3,23 +3,9 @@ Utilities.
 
 A library with useful functions.
 """
-import re
-
 import VLE.factory as factory
 from VLE.models import Comment, Entry, Node, PresetNode, Template
 from VLE.utils.error_handling import VLEMissingRequiredKey, VLEParamWrongType
-
-
-# from https://stackoverflow.com/q/25121165/5173684
-def strip_script_tags(page_source):
-    if page_source is None or page_source == '':
-        return page_source
-    pattern = re.compile(r'\s?on\w+="[^"]+"\s?')
-    pattern = re.compile(r"\s?on\w+='[^']+'\s?")
-    result = re.sub(pattern, "", page_source)
-    pattern2 = re.compile(r'<script[\s\S]+?/script>')
-    result = re.sub(pattern2, "", result)
-    return result
 
 
 # START: API-POST functions
